+++ tests/parser/test_utils.py
@@ -1,8 +1,8 @@
 import unittest
 from datetime import datetime
 
-from ogn.aprs_utils import dmsToDeg, createTimestamp, create_aprs_login
-from ogn.exceptions import AmbigousTimeError
+from ogn.parser.utils import dmsToDeg, createTimestamp
+from ogn.parser.exceptions import AmbigousTimeError
 
 
 class TestStringMethods(unittest.TestCase):
@@ -21,13 +21,6 @@ class TestStringMethods(unittest.TestCase):
     def test_createTimestamp_big_difference(self):
         with self.assertRaises(AmbigousTimeError):
             createTimestamp('123456', reference=datetime(2015, 10, 15, 23, 59, 59))
-
-    def test_create_aprs_login(self):
-        basic_login = create_aprs_login('klaus', -1, 'myApp', '0.1')
-        self.assertEqual('user klaus pass -1 vers myApp 0.1\n', basic_login)
-
-        login_with_filter = create_aprs_login('klaus', -1, 'myApp', '0.1', 'r/48.0/11.0/100')
-        self.assertEqual('user klaus pass -1 vers myApp 0.1 filter r/48.0/11.0/100\n', login_with_filter)
 
 
 if __name__ == '__main__':